backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Supabase
supabase==2.3.4
python-jose[cryptography]==3.3.0

# Pydantic & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI - Google Gemini
google-generativeai==0.3.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Logging & Monitoring
structlog==24.1.0

# Security
passlib[bcrypt]==1.7.4
